User Creation with Email Validator
Created new rest controller Regirstion with
POST api/registration RegistrationRequest as register
GET api/registration/confirm?token=
as token confirmator.
