+++ src/main/java/com/utn/frlp/tecle/service/impl/UserServiceImpl.java
@@ -76,7 +76,7 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
         confirmationTokenService.saveConfirmationToken(confirmationToken);
 
-        return "it works";
+        return token;
     }
 
     @Override
